Preliminary Help window (implementing VScrollbox)

+++ Source/TrackViewer/PopupWindows/QuitWindow.cs
@@ -43,7 +43,7 @@ namespace Orts.TrackViewer.PopupWindows
                 throw new ArgumentNullException(nameof(layout));
 
             layout = base.Layout(layout, 1.5f);
-            quitButton = new Label(this, layout.RemainingWidth / 2, (int)(Owner.TextFontDefault.Height), CatalogManager.Catalog.GetString($"Quit ({InputSettings.UserCommands[UserCommand.QuitWindow].ToString().Max(3)})"), HorizontalAlignment.Center);
+            quitButton = new Label(this, layout.RemainingWidth / 2, Owner.TextFontDefault.Height, CatalogManager.Catalog.GetString($"Quit ({InputSettings.UserCommands[UserCommand.QuitWindow].ToString().Max(3)})"), HorizontalAlignment.Center);
             quitButton.OnClick += QuitButton_OnClick;
             cancelButton = new Label(this, layout.RemainingWidth / 2, Owner.TextFontDefault.Height, CatalogManager.Catalog.GetString($"Cancel ({InputSettings.UserCommands[UserCommand.Cancel].ToString().Max(3)})"), HorizontalAlignment.Center);
             cancelButton.OnClick += CancelButton_OnClick;
